+++ restaurant_finder/templates/index.html
@@ -1,0 +1,191 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>レストラン検索</title>
+  <!-- Tailwind CSS -->
+  <link
+    href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css"
+    rel="stylesheet"
+  />
+  <!-- Leaflet CSS -->
+  <link
+    rel="stylesheet"
+    href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
+  />
+  <!-- Font Awesome -->
+  <link
+    href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+    rel="stylesheet"
+  />
+
+  <style>
+    /* マップの高さ調整 */
+    #map {
+      height: 80vh;
+      transition: all 0.3s ease;
+    }
+    /* ローディングアニメーション */
+    .loading {
+      animation: spin 1s linear infinite;
+    }
+    @keyframes spin {
+      0% {
+        transform: rotate(0deg);
+      }
+      100% {
+        transform: rotate(360deg);
+      }
+    }
+    /* カードホバー */
+    .card {
+      transition: all 0.3s ease;
+      height: 100%;
+    }
+    .card:hover {
+      transform: translateY(-5px);
+      box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
+    }
+    /* 現在地ボタンにパルスアニメーション */
+    .location-pulse {
+      animation: pulse 2s infinite;
+    }
+    @keyframes pulse {
+      0% {
+        box-shadow: 0 0 0 0 rgba(59, 130, 246, 0.5);
+      }
+      70% {
+        box-shadow: 0 0 0 20px rgba(59, 130, 246, 0);
+      }
+      100% {
+        box-shadow: 0 0 0 0 rgba(59, 130, 246, 0);
+      }
+    }
+    /* ピンボタンのツールチップ */
+    .pin-tooltip {
+      pointer-events: none;
+      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
+        0 2px 4px -1px rgba(0, 0, 0, 0.06);
+    }
+    #toggleClickMode:hover .pin-tooltip {
+      opacity: 1;
+      visibility: visible;
+      transform: translate(-50%, -2px);
+    }
+    /* クリックモードがアクティブの場合のアニメーション */
+    .click-mode-active {
+      animation: pulse-border 2s infinite;
+    }
+    @keyframes pulse-border {
+      0% {
+        box-shadow: 0 0 0 0 rgba(245, 158, 11, 0.4);
+      }
+      70% {
+        box-shadow: 0 0 0 10px rgba(245, 158, 11, 0);
+      }
+      100% {
+        box-shadow: 0 0 0 0 rgba(245, 158, 11, 0);
+      }
+    }
+    /* サークル表示 */
+    .leaflet-circle-radius {
+      stroke: #3b82f6;
+      stroke-width: 2;
+      fill: #3b82f6;
+      fill-opacity: 0.1;
+    }
+  </style>
+</head>
+<body class="bg-gray-100 min-h-screen text-base md:text-lg">
+  <!-- ナビゲーションバー -->
+  <nav class="bg-white shadow-lg fixed top-0 w-full z-50">
+    <div class="container mx-auto px-6 py-4">
+      <h1 class="text-3xl font-bold text-gray-800 flex items-center">
+        <i class="fas fa-utensils text-blue-500 mr-2"></i>
+        レストラン検索
+      </h1>
+    </div>
+  </nav>
+
+  <!-- コンテンツ -->
+  <div class="container mx-auto px-4 pt-20 pb-8">
+    <div class="bg-white rounded-xl shadow-lg overflow-hidden">
+      <div class="p-6">
+        <div class="flex flex-wrap gap-4 justify-between items-center mb-6">
+          <!-- ボタン類 -->
+          <div class="flex gap-4">
+            <button
+              id="getCurrentLocation"
+              class="bg-blue-500 hover:bg-blue-600 text-white font-semibold py-3 px-6 rounded-lg flex items-center transition-all duration-300 location-pulse"
+            >
+              <i class="fas fa-location-dot mr-2"></i>
+              現在地を取得
+            </button>
+            <button
+              id="toggleClickMode"
+              class="relative bg-green-500 hover:bg-green-600 text-white font-semibold py-3 px-6 rounded-lg flex items-center transition-all duration-300"
+            >
+              <span class="flex items-center">
+                <i class="fas fa-map-pin mr-2"></i>
+                <span class="pin-text">ピンを追加</span>
+              </span>
+              <div
+                class="pin-tooltip opacity-0 invisible absolute left-1/2 bottom-full mb-2 px-3 py-2 bg-gray-900 text-white text-sm rounded-lg transform -translate-x-1/2 transition-all duration-200 whitespace-nowrap z-50"
+              >
+                地図をクリックして検索位置を設定
+                <div
+                  class="absolute left-1/2 bottom-0 transform -translate-x-1/2 translate-y-full"
+                >
+                  <div class="border-8 border-transparent border-t-gray-900"></div>
+                </div>
+              </div>
+            </button>
+          </div>
+
+          <!-- 検索半径 -->
+          <div class="flex items-center gap-4">
+            <label class="text-gray-700 font-medium text-lg">検索半径:</label>
+            <select
+              id="searchRadius"
+              class="bg-gray-50 border border-gray-300 text-gray-900 rounded-lg py-2.5 px-4 focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all duration-300"
+            >
+              <option value="1000">1km</option>
+              <option value="1500" selected>1.5km</option>
+              <option value="2000">2km</option>
+              <option value="3000">3km</option>
+            </select>
+          </div>
+        </div>
+      </div>
+
+      <!-- マップ -->
+      <div id="map" class="border-t border-b border-gray-200"></div>
+
+      <!-- ローディング指標 -->
+      <div id="loadingIndicator" class="hidden p-4">
+        <div class="flex justify-center items-center">
+          <div class="loading">
+            <i class="fas fa-circle-notch text-4xl text-blue-500"></i>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <!-- 検索結果 -->
+    <div class="mt-8 mb-4">
+      <h2 class="text-2xl font-semibold text-gray-800 mb-2">検索結果</h2>
+      <p id="resultCount" class="text-gray-600"></p>
+    </div>
+
+    <div id="restaurantList" class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+      <!-- レストラン情報がここに動的に追加されます -->
+    </div>
+  </div>
+
+  <!-- Leaflet JS -->
+  <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
+
+  <script src="{{ url_for('static', filename='js/main.js') }}"></script>
+</body>
+</html>
