+++ restaurant_finder/utils.py
@@ -1,0 +1,18 @@
+import math
+
+
+def calculate_distance(lat1, lng1, lat2, lng2):
+    """Return the straight-line distance in meters between two lat/lng pairs."""
+    radius = 6371000
+    lat1_rad = math.radians(lat1)
+    lng1_rad = math.radians(lng1)
+    lat2_rad = math.radians(lat2)
+    lng2_rad = math.radians(lng2)
+
+    dlat = lat2_rad - lat1_rad
+    dlng = lng2_rad - lng1_rad
+
+    a = math.sin(dlat / 2) ** 2 + math.cos(lat1_rad) * math.cos(lat2_rad) * math.sin(dlng / 2) ** 2
+    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))
+
+    return radius * c
